Split into components
No change in functionality.
